Update patch set 16
Patch Set 16: Code-Review-2

(1 comment)

Why are we repeating the same `rootwrap.tpl` and Oslo `sed` commands which already went through many rounds of review and was approved and merged? This goes against the basic principles of DRY. We should not be repeating the maintenance of this same code across *all* the utility containers, as that is not desirable nor sustainable. Code-level changes belong in the container itself where it can be reused across all the utility charts, not in each chart where it has to be repeated.

Patch-set: 16
Reviewer: Gerrit User 1012094 <1012094@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-2